# OTLP to directly writes to storage.Appender