--- a/web/api/v1/api.go
+++ b/web/api/v1/api.go
@@ -316,6 +316,7 @@ func NewAPI(
 			ConvertDelta:            otlpDeltaToCumulative,
 			NativeDelta:             otlpNativeDeltaIngestion,
 			LookbackDelta:           lookbackDelta,
+			IngestCTZeroSample:      ctZeroIngestionEnabled,
 			EnableTypeAndUnitLabels: enableTypeAndUnitLabels,
 		})
 	}