NOTE: continued in #16951 .

Remove intermediary OTLP -> PRW 1.0 step before writing to the appender.  It also adds a CombinedAppender interface, which consolidates error handling, and may make it possible to re-use the library in Mimir, which currently uses it to translate to PRW 1.0.

Similar to https://github.com/prometheus/prometheus/pull/16827, but maintains existing code structure and unit tests.

Alternative to https://github.com/prometheus/prometheus/pull/16784.

Notable Changes:

* Add OTLP-specific metrics for metadata and exemplar write failures:
  * `otlp_without_metadata_appended_samples_total` instead of `remote_write_without_metadata_appended_samples_total`.
  * `otlp_out_of_order_exemplars_total` replaces "Error on ingesting out-of-order exemplars" warning log.
* It now calls `appender.UpdateMetadata` with metadata. The unit is converted using the `otlptranslator` library.
* It now calls `appender.AppendHistogramCTZeroSample` and `appender.AppendCTZeroSample` when the created-timestamp-zero-ingestion feature is enabled.

TODO:

* Unit tests for metadata in write_test.go
* Unit tests for zero timestamp insertion in write_test.go

Benchmark summary: 36% less CPU, 38% fewer bytes, 19% fewer allocations

```
$ benchstat main.txt directappender.txt
goos: linux
goarch: amd64
pkg: github.com/prometheus/prometheus/storage/remote/otlptranslator/prometheusremotewrite
cpu: Intel(R) Xeon(R) CPU @ 2.20GHz
                                                                                                                                                              │    main.txt    │         directappender.txt          │
                                                                                                                                                              │     sec/op     │    sec/op     vs base               │
ConvertBucketLayout/gap_0-2                                                                                                                                      18.95µ ±   8%   18.19µ ±  4%        ~ (p=0.093 n=6)
ConvertBucketLayout/gap_1-2                                                                                                                                      20.38µ ±  11%   18.13µ ±  5%  -11.07% (p=0.009 n=6)
ConvertBucketLayout/gap_2-2                                                                                                                                      19.05µ ± 140%   18.22µ ± 25%        ~ (p=0.699 n=6)
ConvertBucketLayout/gap_3-2                                                                                                                                      23.75µ ±   4%   10.46µ ±  4%  -55.98% (p=0.002 n=6)
ConvertHistogramBucketsToNHCBLayout/gap_0-2                                                                                                                      14.83µ ±   4%   14.75µ ±  4%        ~ (p=0.699 n=6)
ConvertHistogramBucketsToNHCBLayout/gap_1-2                                                                                                                      14.56µ ±   2%   14.53µ ±  3%        ~ (p=0.784 n=6)
ConvertHistogramBucketsToNHCBLayout/gap_2-2                                                                                                                      14.86µ ±  20%   14.66µ ±  9%        ~ (p=0.132 n=6)
ConvertHistogramBucketsToNHCBLayout/gap_3-2                                                                                                                     19.582µ ±  15%   6.852µ ±  2%  -65.01% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_0-2          18.60m ±   8%   14.76m ±  4%  -20.65% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_5-2          30.10m ±  40%   21.15m ±  5%  -29.75% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_10-2         37.82m ±   6%   27.43m ±  4%  -27.48% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_0-2         71.64m ±  13%   48.32m ±  9%  -32.55% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_5-2         70.99m ±   8%   52.67m ± 18%  -25.81% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_10-2        80.03m ±   3%   61.63m ±  2%  -22.99% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_2/exemplars_per_series:_0-2          25.52m ±   5%   14.77m ± 11%  -42.13% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_2/exemplars_per_series:_5-2          31.44m ±   4%   18.84m ±  3%  -40.08% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_2/exemplars_per_series:_10-2         37.92m ±  23%   23.00m ±  4%  -39.35% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_20/exemplars_per_series:_0-2         75.28m ±   4%   30.67m ±  3%  -59.26% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_20/exemplars_per_series:_5-2         85.59m ±   8%   36.87m ±  3%  -56.92% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_20/exemplars_per_series:_10-2        88.03m ±   7%   41.00m ±  1%  -53.43% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_0-2       43.63m ±   4%   29.28m ±  4%  -32.89% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_5-2       65.93m ±  14%   40.59m ±  3%  -38.44% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_10-2      79.91m ±  18%   51.47m ±  3%  -35.59% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_0-2     146.47m ±  24%   79.46m ±  2%  -45.75% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_5-2     170.66m ±  20%   93.22m ±  8%  -45.37% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_10-2     174.0m ±   7%   102.5m ±  4%  -41.09% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_5/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_0-2          18.06m ±   2%   14.80m ±  2%  -18.03% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_5/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_5-2          28.28m ±  16%   21.53m ±  3%  -23.86% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_5/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_10-2         36.69m ±   3%   28.42m ±  3%  -22.54% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_5/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_0-2         62.61m ±   4%   47.47m ±  2%  -24.18% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_5/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_5-2         70.21m ±   4%   52.83m ±  2%  -24.76% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_5/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_10-2        84.62m ±   3%   62.64m ±  1%  -25.98% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_5/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_2/exemplars_per_series:_0-2          24.11m ±   4%   14.76m ±  4%  -38.77% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_5/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_2/exemplars_per_series:_5-2          31.24m ±   3%   19.51m ±  4%  -37.54% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_5/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_2/exemplars_per_series:_10-2         38.30m ±   4%   23.40m ±  2%  -38.92% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_5/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_20/exemplars_per_series:_0-2         75.22m ±   4%   31.72m ±  4%  -57.82% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_5/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_20/exemplars_per_series:_5-2         78.43m ±   3%   36.82m ±  4%  -53.05% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_5/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_20/exemplars_per_series:_10-2        84.26m ±   8%   41.01m ±  1%  -51.33% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_5/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_0-2       43.09m ±   3%   29.98m ±  4%  -30.44% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_5/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_5-2       64.92m ±   3%   40.34m ±  3%  -37.87% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_5/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_10-2      79.51m ±  11%   51.82m ±  2%  -34.82% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_5/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_0-2     139.29m ±   7%   79.55m ±  2%  -42.89% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_5/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_5-2     171.15m ±   7%   93.60m ±  3%  -45.31% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_5/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_10-2     173.2m ±   3%   102.0m ±  2%  -41.07% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_50/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_0-2         18.84m ±   2%   15.49m ±  3%  -17.81% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_50/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_5-2         29.38m ±   6%   22.33m ±  2%  -24.00% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_50/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_10-2        37.53m ±   6%   28.41m ±  8%  -24.31% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_50/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_0-2        60.82m ±   9%   47.81m ±  2%  -21.40% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_50/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_5-2        71.37m ±   5%   52.80m ± 10%  -26.02% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_50/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_10-2       86.75m ±  24%   62.25m ±  2%  -28.24% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_50/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_2/exemplars_per_series:_0-2         24.07m ±   2%   14.97m ±  5%  -37.81% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_50/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_2/exemplars_per_series:_5-2         31.65m ±   7%   19.71m ± 13%  -37.71% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_50/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_2/exemplars_per_series:_10-2        38.34m ±   3%   23.55m ±  2%  -38.58% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_50/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_20/exemplars_per_series:_0-2        74.40m ±   4%   30.92m ±  4%  -58.44% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_50/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_20/exemplars_per_series:_5-2        79.92m ±   4%   37.18m ±  5%  -53.48% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_50/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_20/exemplars_per_series:_10-2       83.36m ±   9%   41.37m ±  3%  -50.38% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_50/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_0-2      44.37m ±   8%   30.61m ±  4%  -31.01% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_50/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_5-2      65.65m ±   2%   40.90m ±  3%  -37.69% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_50/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_10-2     84.19m ±  14%   53.21m ±  3%  -36.81% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_50/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_0-2    135.35m ±   3%   81.49m ±  7%  -39.80% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_50/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_5-2    171.30m ±  10%   94.12m ±  8%  -45.06% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_50/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_10-2    174.4m ±   6%   101.4m ±  2%  -41.85% (p=0.002 n=6)
geomean                                                                                                                                                          21.19m          13.54m        -36.12%

                                                                                                                                                              │   main.txt    │         directappender.txt          │
                                                                                                                                                              │     B/op      │     B/op      vs base               │
ConvertBucketLayout/gap_0-2                                                                                                                                      32.66Ki ± 0%   32.62Ki ± 0%   -0.12% (p=0.002 n=6)
ConvertBucketLayout/gap_1-2                                                                                                                                      32.66Ki ± 0%   32.62Ki ± 0%   -0.12% (p=0.002 n=6)
ConvertBucketLayout/gap_2-2                                                                                                                                      32.66Ki ± 0%   32.62Ki ± 0%   -0.12% (p=0.002 n=6)
ConvertBucketLayout/gap_3-2                                                                                                                                      35.65Ki ± 0%   15.98Ki ± 0%  -55.16% (p=0.002 n=6)
ConvertHistogramBucketsToNHCBLayout/gap_0-2                                                                                                                      24.66Ki ± 0%   24.62Ki ± 0%   -0.16% (p=0.002 n=6)
ConvertHistogramBucketsToNHCBLayout/gap_1-2                                                                                                                      24.66Ki ± 0%   24.62Ki ± 0%   -0.16% (p=0.002 n=6)
ConvertHistogramBucketsToNHCBLayout/gap_2-2                                                                                                                      24.66Ki ± 0%   24.62Ki ± 0%   -0.16% (p=0.002 n=6)
ConvertHistogramBucketsToNHCBLayout/gap_3-2                                                                                                                     27.648Ki ± 0%   7.984Ki ± 0%  -71.12% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_0-2          3.017Mi ± 0%   2.417Mi ± 0%  -19.90% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_5-2          7.235Mi ± 0%   4.486Mi ± 0%  -37.99% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_10-2        11.450Mi ± 0%   6.565Mi ± 0%  -42.66% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_0-2        10.948Mi ± 0%   5.782Mi ± 0%  -47.19% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_5-2        15.168Mi ± 0%   7.868Mi ± 0%  -48.12% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_10-2       19.374Mi ± 0%   9.944Mi ± 0%  -48.67% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_2/exemplars_per_series:_0-2          7.766Mi ± 0%   6.133Mi ± 0%  -21.03% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_2/exemplars_per_series:_5-2         10.644Mi ± 0%   7.757Mi ± 0%  -27.13% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_2/exemplars_per_series:_10-2        13.614Mi ± 0%   9.463Mi ± 0%  -30.49% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_20/exemplars_per_series:_0-2         22.19Mi ± 0%   11.05Mi ± 0%  -50.21% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_20/exemplars_per_series:_5-2         25.07Mi ± 0%   12.67Mi ± 0%  -49.47% (p=0.002 n=6)
…