@krajorama 2025-07-15
When adopting to mimir-prometheus, I've ran into a problem where I have many conflicts with the customizations done in  mimir-prometheus. Turns out most of it will go away with this PR, since we can hide behind interface, so I'd like to throw them away and resync the storage/remote/otlptranslator to upstream, but at which point? 

In the mean time #16842 came in from 3.5 release branch, which is why this PR is also conflicting. 

Was it intentional to have #16842 on main @ArthurSens  ? I don't want to start before this is sorted out and maybe this PR #16855 is rebased @dashpole  ?