@krajorama 2025-07-15
> > Was it intentional to have #16842 on main @ArthurSens ? I don't want to start before this is sorted out and maybe this PR #16855 is rebased @dashpole ?
> 
> Yes, it was intentional. We probably will revert the revert now, but there were some unfortunate misalignments on the spec while 3.5 was being released, so we removed the feature. We'll get it back now

Ack, so we don't need to rebase this PR now and I can essentially sync with upstream at the same base as this PR. And we'll just wait for the revert-revert to move forward.