@dashpole 2025-07-21
It should be easy enough to add the storage.SeriesRef.

Were you able to measure the performance impact of #2?  Would using the `slicelabels` implementation of labels mitigate that?  Or maybe it will just make other things slower?