@krajorama 2025-07-22
> > It should be easy enough to add the storage.SeriesRef.
> 
> 👍
> Anyway, I'll add seriesref on my branch and try to do some measurements.

Well, this turned out to not that easy actually. Nothing in the API says that data points for a metric share the same attributes, so I would have to remember the labels and compare to the previous labels to know if I can reuse the reference and also we turn histogram data points into multiple series so keeping track of those is even more complicated. It's much simpler, what you did to keep track of hashes (and later conflict overflow). I'll do the same instead and do a map of labels hash -> position in the write request.