@krajorama 2025-07-28
I've made a commit (in mimir-prometheus currently) to basically reuse slice labels for sure in the converter: https://github.com/grafana/mimir-prometheus/commit/7f9dd1e244c97e769b2378fde431f82326bf2738 .

With that I get pretty much no change in benchmark from main to the Mimir POC https://github.com/grafana/mimir/pull/12152 .  Even though it only deals with series labels, not exemplar labels. Series labels are passed directly which I can easily manipulate, but Exemplar labels are passed in the [Exemplar type](https://github.com/prometheus/prometheus/blob/2c04f2d7b1a7bbd260ffae70d1bd57b20c0c0d4f/model/exemplar/exemplar.go#L26), so that would be way harder.

If you'd be ok with these tricks to preserve slice labels in the interface and only switch to the real labels when needed, this should work. According to BenchmarkPrometheusConverter_FromMetrics in Prometheus, this trick has no overhead - the scratchbuilder/builder already deals with slices anyway, converting to stringlabels/dedupelabels is anyway the last step.