@krajorama 2025-07-28
> I'm OK with having our own labels implementation if needed. I suspect we can simplify the interface for our labels implementation (probably don't need separate scratchbuilder and builder types, for example). It might be better to copy the separators we need, rather than make them part of the public package for labels.

Yes, we can refine it further for sure, I wanted to get to something working soon. There might be more possibilities to optimize as well.

> 
> Thoughts on what the overall package should be named? `prometheusremotewrite` doesn't make sense. I actually like `otlpappender` for the name.

Also agree, doesn't make much sense to call it `prometheusremotewrite` anymore. `otlpappender` sounds good.