@dashpole 2025-07-29
rebased, but need to address comments