@krajorama 2025-07-30
The two TODOs has been addressed in #16951:

    Unit tests for metadata in write_test.go
    Unit tests for zero timestamp insertion in write_test.go